Assignment 4/requirements.txt
openai
python-docx
pandas
requests
beautifulsoup4
PyPDF2
python-dotenv
